Add TraineeAttributes degree_attributes validation
* Validate for presence
